// src/types/jumpseller.ts
export interface JumpSellerRequest {

  x_url_callback: string;
  x_url_complete: string;
  x_url_cancel: string;
  x_account_id: string;
  x_amount: string;
  x_currency: string;
  x_reference: string;
  x_shop_country: string;
  x_shop_name: string;
  x_description: string;
  x_customer_first_name: string;
  x_customer_last_name: string;
  x_customer_email: string;
  x_customer_phone: string;
  x_customer_shipping_first_name: string;
  x_customer_shipping_last_name: string;
  x_customer_shipping_city: string;
  x_customer_shipping_address1: string;
  x_customer_shipping_address2: string;
  x_customer_shipping_state: string;
  x_customer_shipping_zip: string;
  x_customer_shipping_country: string;
  x_customer_shipping_phone: string;
  x_customer_billing_first_name: string;
  x_customer_billing_last_name: string;
  x_customer_billing_city: string;
  x_customer_billing_address1: string;
  x_customer_billing_address2: string;
  x_customer_billing_state: string;
  x_customer_billing_zip: string;
  x_customer_billing_country: string;
  x_customer_billing_phone: string;
  x_customer_taxid: string;
  x_signature: string;


}